dbs: move to `internal/dbs`
