+++ cmd/sipsctl/cmd/migrate.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/DeedleFake/sips/dbs"
-	"github.com/DeedleFake/sips/dbs/migrate"
+	"github.com/DeedleFake/sips/internal/dbs"
+	"github.com/DeedleFake/sips/internal/dbs/migrate"
 	"github.com/spf13/cobra"
 )
