+++ internal/dbs/migrate/2021-06-22T15:10:17_AddPinStatus.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/DeedleFake/sips"
-	"github.com/DeedleFake/sips/dbs"
+	"github.com/DeedleFake/sips/internal/dbs"
 	"github.com/asdine/storm"
 )
