+++ internal/dbs/migrate/generate.bash
